STREAMLIT/requirements.txt
numpy
pandas
requests
streamlit
plotly.express
Pillow
sqlalchemy==1.4.40